+++ CHANGELOG.md
@@ -13,12 +13,16 @@ This project adheres to [Semantic Versioning], with the exception that minor rel
 
 - ✨ Add version manifest ([#79], [#97]) ([**@denialhaag**], [**@burgholzer**])
 
+### Removed
+
+- 🔥 Remove support for Debug builds on Windows ([#106]) ([**@denialhaag**])
+
 ## [1.1.0] - 2026-01-07
 
 ### Added
 
 - ✨ Add support for distributed zstd binaries so that installer scripts and GitHub Action now automatically download and use platform-specific zstd binaries for decompression ([#61]) ([**@burgholzer**])
-- ✨ Add support for debug builds on Windows with new `use_debug` option available in both installer scripts and GitHub Action ([#61]) ([**@burgholzer**])
+- ✨ Add support for Debug builds on Windows with new `use_debug` option available in both installer scripts and GitHub Action ([#61]) ([**@burgholzer**])
 - ✅ Add comprehensive integration tests that exercise actual download and extraction paths ([#61]) ([**@burgholzer**])
 
 ### Changed
@@ -61,6 +65,7 @@ _This is the initial release of the `setup-mlir` project._
 
 <!-- PR links -->
 
+[#106]: https://github.com/munich-quantum-software/setup-mlir/pull/106
 [#97]: https://github.com/munich-quantum-software/setup-mlir/pull/97
 [#79]: https://github.com/munich-quantum-software/setup-mlir/pull/79
 [#61]: https://github.com/munich-quantum-software/setup-mlir/pull/61